[IMPLICITS] Removed some implicit conversion constraints.

+++ modules/api/src/test/scala/org/platanios/tensorflow/api/ops/CallbackSuite.scala
@@ -30,45 +30,45 @@ import org.scalatest.junit.JUnitSuite
   * @author Emmanouil Antonios Platanios
   */
 class CallbackSuite extends JUnitSuite {
-  def square(input: Tensor[Float]): Tensor[Float] = {
+  def square(input: Tensor[Double]): Tensor[Double] = {
     input.square
   }
 
-  def add(inputs: Seq[Tensor[Float]]): Tensor[Float] = {
+  def add(inputs: Seq[Tensor[Double]]): Tensor[Double] = {
     TensorMath.addN(inputs)
   }
 
   @Test def testIdentitySingleInputSingleOutputCallback(): Unit = using(Graph()) { graph =>
     val (input, output) = Op.createWith(graph) {
-      val input = Basic.placeholder[Float]()
-      val output = Callback.callback(square, input, FLOAT32)
+      val input = Basic.placeholder[Double]()
+      val output = Callback.callback(square, input, FLOAT64)
       (input, output)
     }
     val session = Session(graph = graph)
-    val outputValue = session.run(Map(input -> Tensor(2f, 5f)), output)
-    assert(outputValue.dataType == FLOAT32)
+    val outputValue = session.run(Map(input -> Tensor(2.0, 5.0)), output)
+    assert(outputValue.dataType == FLOAT64)
     assert(outputValue.shape == Shape(2))
-    assert(outputValue(0).scalar == 4f)
-    assert(outputValue(1).scalar == 25f)
+    assert(outputValue(0).scalar == 4.0)
+    assert(outputValue(1).scalar == 25.0)
   }
 
   @Test def testIdentityMultipleInputSingleOutputCallback(): Unit = using(Graph()) { graph =>
     val (input1, input2, input3, output) = Op.createWith(graph) {
-      val input1 = Basic.placeholder[Float]()
-      val input2 = Basic.placeholder[Float]()
-      val input3 = Basic.placeholder[Float]()
-      val output = Callback.callback(add, Seq(input1, input2, input3), FLOAT32)
+      val input1 = Basic.placeholder[Double]()
+      val input2 = Basic.placeholder[Double]()
+      val input3 = Basic.placeholder[Double]()
+      val output = Callback.callback(add, Seq(input1, input2, input3), FLOAT64)
       (input1, input2, input3, output)
     }
     val session = Session(graph = graph)
     val outputValue = session.run(Map(
-      input1 -> Tensor(2.0f, 5.0f),
-      input2 -> Tensor(-1.3f, 3.1f),
-      input3 -> Tensor(8.9f, -4.1f)
+      input1 -> Tensor(2.0, 5.0),
+      input2 -> Tensor(-1.3, 3.1),
+      input3 -> Tensor(8.9, -4.1)
     ), output)
-    assert(outputValue.dataType == FLOAT32)
+    assert(outputValue.dataType == FLOAT64)
     assert(outputValue.shape == Shape(2))
-    assert(outputValue(0).scalar == 9.6f)
-    assert(outputValue(1).scalar == 4.0f)
+    assert(outputValue(0).scalar == 9.6)
+    assert(outputValue(1).scalar == 4.0)
   }
 }
